PR comments * Split the TabBar and SimpleTabBar definitions into their own file * Cleaned up how arrow key navigation renders its children to remove the need for a key at this level * Tidied up how tables render no content
