+++ src/components/TopBar.js
@@ -331,9 +331,7 @@ const TopBar = ({ showMenu = true, title, href, children }) => {
   }
   const mainRef = useRef()
 
-  return div({
-    role: 'banner'
-  }, [
+  return div({ role: 'banner' }, [
     h(SkipNavLink, { ref: mainRef }),
     h(Transition, {
       in: navShown,
