+++ src/pages/workflows/List.js
@@ -3,9 +3,10 @@ import * as qs from 'qs'
 import { useState } from 'react'
 import { div, h } from 'react-hyperscript-helpers'
 import { AutoSizer } from 'react-virtualized'
-import { Link, TabBar } from 'src/components/common'
+import { Link } from 'src/components/common'
 import FooterWrapper from 'src/components/FooterWrapper'
 import { DelayedSearchInput } from 'src/components/input'
+import { TabBar } from 'src/components/TabBar'
 import { FlexTable, HeaderCell, Sortable, TooltipCell } from 'src/components/table'
 import TopBar from 'src/components/TopBar'
 import { Ajax } from 'src/libs/ajax'
