+++ src/pages/workflows/workflow/WorkflowDetails.js
@@ -5,10 +5,11 @@ import { Fragment, useState } from 'react'
 import { div, h, h2, label, span } from 'react-hyperscript-helpers'
 import { AutoSizer } from 'react-virtualized'
 import * as breadcrumbs from 'src/components/breadcrumbs'
-import { ButtonSecondary, IdContainer, Link, Select, TabBar } from 'src/components/common'
+import { ButtonSecondary, IdContainer, Link, Select } from 'src/components/common'
 import FooterWrapper from 'src/components/FooterWrapper'
 import { centeredSpinner, icon } from 'src/components/icons'
 import { MarkdownViewer, newWindowLinkRenderer } from 'src/components/markdown'
+import { TabBar } from 'src/components/TabBar'
 import { FlexTable, HeaderCell } from 'src/components/table'
 import TooltipTrigger from 'src/components/TooltipTrigger'
 import TopBar from 'src/components/TopBar'
